Use Paginator.wrap to implement table
Rather than implementing `Paginator.wrap` to refer to decorators built
for `PaginatorTable`, move the implementation to `Paginator.wrap`. The
table is then built with calls to `Paginator.wrap`.

Additionally, because it now makes more sense there, move the
`has_paginator` decorator to the same module as the base paginator,
not the table.

The annotation for `BaseClient.paginated: PaginatorTable` was added as
part of the prior implementation of `Paginator.wrap`. It is no longer
needed, so remove it.
