+++ src/globus_sdk/paging/__init__.py
@@ -1,9 +1,9 @@
-from .base import Paginator
+from .base import Paginator, has_paginator
 from .last_key import LastKeyPaginator
 from .limit_offset import HasNextPaginator, LimitOffsetTotalPaginator
 from .marker import MarkerPaginator
 from .next_token import NextTokenPaginator
-from .table import PaginatorTable, has_paginator
+from .table import PaginatorTable
 
 __all__ = (
     "Paginator",
