+++ src/globus_sdk/paging/base.py
@@ -1,13 +1,16 @@
 import abc
+import functools
 import inspect
 from typing import (
     Any,
     Callable,
     Dict,
+    Generic,
     Iterable,
     Iterator,
     List,
     Optional,
+    Type,
     TypeVar,
     cast,
 )
@@ -19,6 +22,15 @@ from globus_sdk.response import GlobusHTTPResponse
 PageT = TypeVar("PageT", bound=GlobusHTTPResponse)
 P = te.ParamSpec("P")
 R = TypeVar("R", bound=GlobusHTTPResponse)
+C = TypeVar("C", bound=Callable[..., GlobusHTTPResponse])
+
+
+# stub for mypy
+class _PaginatedFunc(Generic[PageT]):
+    _has_paginator: bool
+    _paginator_class: Type["Paginator[PageT]"]
+    _paginator_items_key: Optional[str]
+    _paginator_params: Dict[str, Any]
 
 
 class Paginator(Iterable[PageT], metaclass=abc.ABCMeta):
@@ -105,24 +117,68 @@ class Paginator(Iterable[PageT], metaclass=abc.ABCMeta):
         Although the syntax is slightly more verbose, this allows `mypy` and other type
         checkers to more accurately infer the type of the paginator.
         """
-        # these import are deferred to avoid circular dependencies
-        # `globus_sdk.paging` is needed by clients to build paginator tables
-        # but the client class is needed here to be able to check that methods are
-        # methods of clients
-        # the table is needed here for typing, but requires Paginator
-        from ..client import BaseClient
-
         if not inspect.ismethod(method):
             raise TypeError(f"Paginator.wrap can only be used on methods, not {method}")
+        if not getattr(method, "_has_paginator", False):
+            raise ValueError(f"'{method}' is not a paginated method")
 
-        if not isinstance(method.__self__, BaseClient):
-            raise ValueError(
-                "Paginator.wrap can only be used on methods of globus-sdk clients"
+        as_paginated = cast(_PaginatedFunc[PageT], method)
+        paginator_class = as_paginated._paginator_class
+        paginator_params = as_paginated._paginator_params
+        paginator_items_key = as_paginated._paginator_items_key
+
+        @functools.wraps(method)
+        def paginated_method(*args: Any, **kwargs: Any) -> Paginator[PageT]:
+            return paginator_class(
+                method,
+                client_args=list(args),
+                client_kwargs=kwargs,
+                items_key=paginator_items_key,
+                **paginator_params,
             )
 
-        try:
-            ret = getattr(method.__self__.paginated, method.__name__)
-        except AttributeError as e:
-            raise ValueError(f"{method.__name__} is not a paginated method") from e
+        return cast(Callable[P, Paginator[R]], paginated_method)
 
-        return cast(Callable[P, Paginator[R]], ret)
+
+def has_paginator(
+    paginator_class: Type[Paginator[PageT]],
+    items_key: Optional[str] = None,
+    **paginator_params: Any,
+) -> Callable[[C], C]:
+    """
+    Mark a callable -- typically a client method -- as having pagination parameters.
+    Usage:
+
+    >>> class MyClient(BaseClient):
+    >>>     @has_paginator(MarkerPaginator)
+    >>>     def foo(...): ...
+
+    This will mark ``MyClient.foo`` as paginated with marker style pagination.
+    It will then be possible to get a paginator for ``MyClient.foo`` via
+
+    >>> c = MyClient(...)
+    >>> paginator = c.paginated.foo()
+    """
+
+    def decorate(func: C) -> C:
+        as_paginated = cast(_PaginatedFunc[PageT], func)
+        as_paginated._has_paginator = True
+        as_paginated._paginator_class = paginator_class
+        as_paginated._paginator_items_key = items_key
+        as_paginated._paginator_params = paginator_params
+
+        func.__doc__ = f"""{func.__doc__}
+
+        **Paginated Usage**
+
+        This method supports paginated access. To use the paginated variant, give the
+        same arguments as normal, but prefix the method name with ``paginated``, as in
+
+        >>> client.paginated.{func.__name__}(...)
+
+        For more information, see
+        :ref:`how to make paginated calls <making_paginated_calls>`.
+        """
+        return func
+
+    return decorate
